+++ app/controllers/registrants/build_controller.rb
@@ -76,7 +76,7 @@ class Registrants::BuildController < ApplicationController
 
     # If there's an organization_membership record, the USA checking will happen after
     # that is created.
-    @registrant.skip_usa_membership_checking = true if @organization_membership
+    @registrant.skip_organization_membership_checking = true if @organization_membership
 
     @registrant.status = "base_details"
     if @registrant.save
