+++ app/models/organization/iuf.rb
@@ -1,11 +1,15 @@
 module Organization
-  class Iuf
+  class Iuf < None
     def title
       "IUF"
     end
 
     def information_partial
       "iuf"
+    end
+
+    def automated_checking_class
+      MembershipChecker::Iuf
     end
   end
 end
