+++ app/models/organization/usa.rb
@@ -1,11 +1,15 @@
 module Organization
-  class Usa
+  class Usa < None
     def title
       "USA"
     end
 
     def information_partial
       "usa"
+    end
+
+    def automated_checking_class
+      MembershipChecker::Usa
     end
   end
 end
