+++ app/views/organization_memberships/_member_row.html.haml
@@ -23,8 +23,8 @@
       - unless reg.organization_membership_confirmed?
         %br
         = reg.organization_membership_system_status
-        - if @config.organization_membership_usa?
-          = link_to "Refresh USA status", refresh_usa_status_organization_membership_path(id: reg.id), remote: true, method: :post
+        - if @config.organization_membership_config.automated_checking?
+          = link_to "Refresh #{@config.organization_membership_config.title} status", refresh_organization_status_organization_membership_path(id: reg.id), remote: true, method: :post
     %td
       - if reg.organization_membership_system_confirmed?
         = "Confirmed by System"
