+++ app/views/registrants/_organization_membership.html.haml
@@ -4,7 +4,7 @@
       %p
         %b= organization_membership_label(@config)
         = organization_membership.member_number
-      - if @config.organization_membership_usa?
+      - if @config.organization_membership_config.automated_checking?
         %p
           %b= OrganizationMembership.human_attribute_name(:system_status)
           = render "/registrants/organization_membership_badge", organization_membership: organization_membership
