+++ app/views/registrants/_organization_membership_status.html.haml
@@ -1,22 +1,22 @@
 .row
   .small-6.columns
-    USA membership status
-  .small-6.column.js--usa_member_status
+    #{organization_config.title} membership status
+  .small-6.column.js--organization_member_status
     = render "/registrants/organization_membership_badge", organization_membership: organization_membership
     = link_to organization_membership_welcome_path, target: "_blank" do
       %span.help_bubble{ title: "More details" }
 
-.row.js--usa_member_status
+.row.js--organization_member_status
   .small-6.columns
-    USA Details
+    #{organization_config.title} Details
   .small-6.columns
     = organization_membership.member_number
     %br
     = organization_membership.system_status
 .row
   .small-12.columns
-    = link_to "Recheck USA Status", refresh_usa_status_registrant_path(organization_membership.registrant.bib_number), remote: true, method: :put, class: "js--usa_member_status_refresh"
-    .js--usa_member_refresh_notice.is--hidden
+    = link_to "Recheck #{organization_config.title} Status", refresh_organization_status_registrant_path(organization_membership.registrant.bib_number), remote: true, method: :put, class: "js--organization_member_status_refresh"
+    .js--organization_member_refresh_notice.is--hidden
       Checking....
       %br
       = link_to "Reload Page", "javascript:window.location.reload()"
