--- app/views/welcome/usa_membership.html.haml
@@ -1,8 +1,0 @@
-%h1 USA Membership
-
-- if @config.organization_membership_usa?
-  = render "usa"
-- else
-  %p
-    This convention is not a "USA" convention
-
