+++ spec/controllers/registrants/build_controller_spec.rb
@@ -8,7 +8,7 @@ describe Registrants::BuildController do
 
     allow_any_instance_of(EventConfiguration).to receive(:registration_closed?).and_return(false)
     allow_any_instance_of(EventConfiguration).to receive(:organization_membership_config?).and_return(true)
-    allow_any_instance_of(EventConfiguration).to receive(:organization_membership_usa?).and_return(true)
+    allow_any_instance_of(EventConfiguration).to receive(:organization_membership_config).and_return(Organization::Usa.new)
   end
 
   # This should return the minimal set of attributes required to create a valid
