+++ spec/workers/update_organization_membership_status_worker_spec.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 
-describe UpdateUsaMembershipStatusWorker do
+describe UpdateOrganizationMembershipStatusWorker do
   subject(:worker) { described_class.new }
 
   let!(:config) { FactoryBot.create(:event_configuration, :with_usa) }
   let(:registrant) { FactoryBot.create(:competitor) }
 
   before do
-    allow_any_instance_of(UsaMembershipChecker).to receive(:current_member?).and_return(true)
+    allow_any_instance_of(MembershipChecker::Usa).to receive(:current_member?).and_return(true)
   end
 
   it "can check a member" do
